pyproject.toml
[tool.poetry]
name = "bfti-bot"
version = "0.1.0"
description = ""
authors = ["coma64 <zaton.tristan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"
coloredlogs = "^15.0"
aiohttp = "^3.7.4"
rope = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
neovim = "^0.3.1"
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py38']
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "88"
